usize

+++ boot/src/main.rs
@@ -30,7 +30,7 @@ fn panic_handler(_info: &core::panic::PanicInfo) -> ! {
     loop {}
 }
 
-fn page_allocator(num_pages: usize) -> Option<u64> {
+fn page_allocator(num_pages: usize) -> Option<usize> {
     let Some(bs) = BOOT_CTX.get_bs() else {
         return None;
     };
@@ -48,10 +48,10 @@ fn page_allocator(num_pages: usize) -> Option<u64> {
         (bs.set_mem)(paddr as *mut core::ffi::c_void, PAGE_SIZE, 0);
     };
 
-    Some(paddr)
+    Some(paddr as usize)
 }
 
-fn prepare_kernel_args(paddr: u64, mem_map: &MemoryMap) -> Option<()> {
+fn prepare_kernel_args(paddr: usize, mem_map: &MemoryMap) -> Option<()> {
     let gop = BOOT_CTX.locate_protocol::<graphics_output::Protocol>(graphics_output::PROTOCOL_GUID)?;
     let mode = unsafe {(*gop).mode.as_ref()}?;
 
@@ -78,7 +78,7 @@ fn prepare_kernel_args(paddr: u64, mem_map: &MemoryMap) -> Option<()> {
 
 fn setup_mem_paging<T>(pt_mgr: &PageTableManager<T>, mem_map: &MemoryMap)  ->  Result<(), efi::Status>
 where
-    T: Fn(usize) -> Option<u64>
+    T: Fn(usize) -> Option<usize>
 {
     let total_mem = mem_map.total_memory;
     let num_1gb_pdpe = (total_mem + (ONE_GB - 1)) / ONE_GB;
@@ -87,11 +87,11 @@ where
     assert!(num_1gb_pdpe < paging::PAGE_TABLE_NUM_ENTRIES);
 
     let mut paddr = PhysicalAddress(0x0);
-    let mut vaddr = VirtualAddress(KERNEL_DS_ADDR as u64);
+    let mut vaddr = VirtualAddress(KERNEL_DS_ADDR);
     for _ in 0..num_1gb_pdpe {
         pt_mgr.map_1gb_page(paddr, vaddr);
-        paddr += ONE_GB as u64;
-        vaddr += ONE_GB as u64;
+        paddr += ONE_GB as usize;
+        vaddr += ONE_GB as usize;
     }
 
     Ok(())
@@ -99,14 +99,14 @@ where
 
 fn setup_kernel_paging<T>(pt_mgr: &PageTableManager<T>, kernel: &Kernel)  ->  Result<(), efi::Status>
 where
-    T: Fn(usize) -> Option<u64>
+    T: Fn(usize) -> Option<usize>
 {
     let mut start_paddr = kernel.kernel_base;
     let mut start_vaddr = kernel.kernel_vaddr;
     for _ in 0..kernel.kernel_pages {
         if pt_mgr.map_page( start_paddr.get_raw(), start_vaddr.get_raw()) {
-            start_vaddr += PAGE_SIZE as u64;
-            start_paddr += PAGE_SIZE as u64;
+            start_vaddr += PAGE_SIZE;
+            start_paddr += PAGE_SIZE;
         } else {
             printer::print("FAILED TO MAP KERNEL\n");
             return Err(efi::Status::OUT_OF_RESOURCES);
@@ -118,16 +118,16 @@ where
 
 fn setup_image_identity_map<T>(h: efi::Handle, pt_mgr: &PageTableManager<T>) -> Result<(), efi::Status>
 where
-    T: Fn(usize) -> Option<u64> {
+    T: Fn(usize) -> Option<usize> {
     let loaded_image =
         BOOT_CTX.handle_protocol::<loaded_image::Protocol>(h, loaded_image::PROTOCOL_GUID).ok_or(efi::Status::PROTOCOL_ERROR)?;
 
     let mut image_base = unsafe {
-        (*loaded_image).image_base as u64
+        (*loaded_image).image_base as usize
     };
 
     let image_size = unsafe {
-        (*loaded_image).image_size as u64
+        (*loaded_image).image_size as usize
     };
 
     let pages = image_size >> 12;
@@ -135,13 +135,13 @@ where
     printer::print(&format!("Mapping image. Base 0x{:x} Size 0x{:x}\n", image_base, image_size));
     for _ in 0..pages {
         pt_mgr.map_page(image_base, image_base);
-        image_base += PAGE_SIZE as u64;
+        image_base += PAGE_SIZE;
     }
     Ok(())
 }
 
 
-fn setup_paging(h: efi::Handle, kernel: &Kernel, mem_map: &MemoryMap) -> Result<PageTableManager<impl Fn(usize) -> Option<u64>>, efi::Status> {
+fn setup_paging(h: efi::Handle, kernel: &Kernel, mem_map: &MemoryMap) -> Result<PageTableManager<impl Fn(usize) -> Option<usize>>, efi::Status> {
     let Some(pt_mgr) = PageTableManager::new(page_allocator) else {
         printer::print("setup_paging failed");
         return  Err(efi::Status::INVALID_PARAMETER);
